refactor to reduce parser file size
