+++ tests/test_source_by_pc.py
@@ -1,5 +1,6 @@
 import unittest
-from solc_json_parser.parser import SolidityAst, get_in
+from solc_json_parser.parser import SolidityAst
+from solc_json_parser.ast_shared import get_in
 from typing import Dict, Any, List
